[ingress-nginx] Impovment isolated mode.
