+++ modules/402-ingress-nginx/crds/doc-ru-ingress-nginx.yaml
@@ -125,6 +125,7 @@ spec:
                     * `IsolatedFilesystem` — конфигурация NGINX проверяется в отдельной файловой системе без доступа к файловой системе работающего сервера NGINX, что эффективно сокращает поверхность атаки. Область проверки ограничена параметром `ingressClass`, что обеспечивает «сильную» валидацию.
                     * `IsolatedProcess` — конфигурация NGINX проверяется внутри изолированной среды (песочницы), где отслеживается каждый системный вызов. Это наиболее безопасный режим изоляции, ограниченный параметром `ingressClass`.
                     > **Внимание.** Некоторые защищенные операционные системы могут блокировать трассировку системных вызовов без дополнительных настроек, поэтому на данный момент этот режим является опциональным.
+                    > **Внимание.** Для режима `IsolatedProcess` требуется `containerd v2` на целевых узлах.
                     > **Внимание.** Данный режим является экспериментальным.
                 annotationValidationEnabled:
                   description: |
@@ -731,6 +732,7 @@ spec:
                     * `IsolatedFilesystem` — конфигурация NGINX проверяется в отдельной файловой системе без доступа к файловой системе работающего сервера NGINX, что эффективно сокращает поверхность атаки. Область проверки ограничена параметром `ingressClass`, что обеспечивает «сильную» валидацию.
                     * `IsolatedProcess` — конфигурация NGINX проверяется внутри изолированной среды (песочницы), где отслеживается каждый системный вызов. Это наиболее безопасный режим изоляции, ограниченный параметром `ingressClass`.
                     > **Внимание.** Некоторые защищенные операционные системы могут блокировать трассировку системных вызовов без дополнительных настроек, поэтому на данный момент этот режим является опциональным.
+                    > **Внимание.** Для режима `IsolatedProcess` требуется `containerd v2` на целевых узлах.
                     > **Внимание.** Данный режим является экспериментальным.
 
                 annotationValidationEnabled:
