+++ modules/402-ingress-nginx/template_tests/controller_test.go
@@ -242,7 +242,7 @@ var _ = Describe("Module :: ingress-nginx :: helm template :: controllers", func
 		checkVersion := func(version string) {
 			hec.ValuesSet("ingressNginx.defaultControllerVersion", version)
 
-			fullMode := render(`
+			isolatedProcess := render(`
 - name: sandbox
   spec:
     ingressClass: nginx
@@ -252,16 +252,18 @@ var _ = Describe("Module :: ingress-nginx :: helm template :: controllers", func
     controllerVersion: "` + version + `"
     validationIsolationMode: IsolatedProcess
 `)
-			Expect(fullMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
-			Expect(fullMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
-			Expect(fullMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
-			Expect(fullMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
-			Expect(fullMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("serviceAccountName: validator-full"))
-			Expect(fullMode["ingress-nginx/templates/validator/deployment.yaml"]).NotTo(ContainSubstring("--watch-namespace=d8-ingress-nginx"))
-			Expect(fullMode["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: validator-full\n  namespace: d8-ingress-nginx"))
-			Expect(fullMode["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: d8:ingress-nginx:validator-full"))
+			Expect(isolatedProcess["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
+			Expect(isolatedProcess["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
+			Expect(isolatedProcess["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("name: net.ipv4.ip_unprivileged_port_start"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("serviceAccountName: validator-full"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/deployment.yaml"]).NotTo(ContainSubstring("--watch-namespace=d8-ingress-nginx"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("name: net.ipv4.ip_unprivileged_port_start"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: validator-full\n  namespace: d8-ingress-nginx"))
+			Expect(isolatedProcess["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: d8:ingress-nginx:validator-full"))
 
-			standardMode := render(`
+			isolatedFilesystem := render(`
 - name: sandbox
   spec:
     ingressClass: nginx
@@ -271,28 +273,30 @@ var _ = Describe("Module :: ingress-nginx :: helm template :: controllers", func
     controllerVersion: "` + version + `"
     validationIsolationMode: IsolatedProcess
 `)
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /chroot/tmp"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/tmp"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/ingress-controller"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/etc/ingress-controller"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/ssl/"))
-			Expect(standardMode["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/etc/nginx/ssl/"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/tmp"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/ssl/"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/webhook-ssl/"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/tmp"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/etc/nginx/ssl/"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/webhook-ssl/"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("serviceAccountName: validator-full"))
-			Expect(standardMode["ingress-nginx/templates/validator/deployment.yaml"]).NotTo(ContainSubstring("--watch-namespace=d8-ingress-nginx"))
-			Expect(standardMode["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: validator-full\n  namespace: d8-ingress-nginx"))
-			Expect(standardMode["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: d8:ingress-nginx:validator-full"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /chroot/tmp"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/tmp"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/ingress-controller"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/etc/ingress-controller"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/ssl/"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/etc/nginx/ssl/"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/controller/controller.yaml"]).To(ContainSubstring("name: net.ipv4.ip_unprivileged_port_start"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("value: IsolatedProcess"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/tmp"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/ssl/"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/webhook-ssl/"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/tmp"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /validation-chroot/etc/nginx/ssl/"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("mountPath: /chroot/etc/nginx/webhook-ssl/"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("serviceAccountName: validator-full"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).NotTo(ContainSubstring("--watch-namespace=d8-ingress-nginx"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("name: net.ipv4.ip_unprivileged_port_start"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: validator-full\n  namespace: d8-ingress-nginx"))
+			Expect(isolatedFilesystem["ingress-nginx/templates/validator/rbac-for-us.yaml"]).To(ContainSubstring("name: d8:ingress-nginx:validator-full"))
 
-			disabledMode := render(`
+			noIsolation := render(`
 - name: sandbox
   spec:
     ingressClass: nginx
@@ -302,10 +306,14 @@ var _ = Describe("Module :: ingress-nginx :: helm template :: controllers", func
     controllerVersion: "` + version + `"
     validationIsolationMode: NoIsolation
 `)
-			Expect(disabledMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("serviceAccountName: validator"))
-			Expect(disabledMode["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("--watch-namespace=d8-ingress-nginx"))
-			Expect(disabledMode["ingress-nginx/templates/validator/rbac-for-us.yaml"]).NotTo(ContainSubstring("name: validator-full"))
-			Expect(disabledMode["ingress-nginx/templates/validator/rbac-for-us.yaml"]).NotTo(ContainSubstring("name: d8:ingress-nginx:validator-full"))
+			Expect(noIsolation["ingress-nginx/templates/controller/controller.yaml"]).NotTo(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
+			Expect(noIsolation["ingress-nginx/templates/validator/deployment.yaml"]).NotTo(ContainSubstring("name: VALIDATION_ISOLATION_MODE"))
+			Expect(noIsolation["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("serviceAccountName: validator"))
+			Expect(noIsolation["ingress-nginx/templates/validator/deployment.yaml"]).To(ContainSubstring("--watch-namespace=d8-ingress-nginx"))
+			Expect(noIsolation["ingress-nginx/templates/validator/rbac-for-us.yaml"]).NotTo(ContainSubstring("name: validator-full"))
+			Expect(noIsolation["ingress-nginx/templates/validator/rbac-for-us.yaml"]).NotTo(ContainSubstring("name: d8:ingress-nginx:validator-full"))
+			Expect(noIsolation["ingress-nginx/templates/controller/controller.yaml"]).NotTo(ContainSubstring("name: net.ipv4.ip_unprivileged_port_start"))
+			Expect(noIsolation["ingress-nginx/templates/validator/deployment.yaml"]).NotTo(ContainSubstring("name: net.ipv4.ip_unprivileged_port_start"))
 		}
 
 		checkVersion("1.14")
