+++ modules/402-ingress-nginx/templates/validation.yaml
@@ -2,6 +2,7 @@
 {{- $kubernetesVersion := .Values.global.discovery.kubernetesVersion }}
 ---
 {{- $policyName := "ingressnginxcontroller-settings.deckhouse.io" }}
+{{- $warningPolicyName := "ingressnginxcontroller-validation-isolation-warning.deckhouse.io" }}
 {{- $inlets := list "LoadBalancer"  "LoadBalancerWithProxyProtocol" "HostPort" "HostPortWithProxyProtocol" "HostWithFailover" }}
 apiVersion: {{ include "helm_lib_get_api_version_by_kind" (list . "ValidatingAdmissionPolicy") }}
 kind: ValidatingAdmissionPolicy
@@ -32,4 +33,31 @@ metadata:
 spec:
   policyName: {{ $policyName }}
   validationActions: [Deny]
+---
+apiVersion: {{ include "helm_lib_get_api_version_by_kind" (list . "ValidatingAdmissionPolicy") }}
+kind: ValidatingAdmissionPolicy
+metadata:
+  name: {{ $warningPolicyName }}
+  {{- include "helm_lib_module_labels" (list . ) | nindent 2 }}
+spec:
+  failurePolicy: Fail
+  matchConstraints:
+    resourceRules:
+      - apiGroups:   ["deckhouse.io"]
+        apiVersions: ["*"]
+        operations:  ["CREATE", "UPDATE"]
+        resources:   ["ingressnginxcontrollers"]
+  validations:
+    - expression: '!(has(object.spec.validationIsolationMode) && object.spec.validationIsolationMode == "IsolatedProcess")'
+      reason: Forbidden
+      messageExpression: '"validationIsolationMode IsolatedProcess requires containerd v2 on target nodes; it may not work on containerd v1 or mixed-runtime clusters. See: https://deckhouse.io/products/kubernetes-platform/documentation/v1/admin/configuration/platform-scaling/node/migrating.html"'
+---
+apiVersion: {{ include "helm_lib_get_api_version_by_kind" (list . "ValidatingAdmissionPolicyBinding") }}
+kind: ValidatingAdmissionPolicyBinding
+metadata:
+  name: {{ $warningPolicyName }}
+  {{- include "helm_lib_module_labels" (list . ) | nindent 2 }}
+spec:
+  policyName: {{ $warningPolicyName }}
+  validationActions: [Warn]
 {{- end }}
