+++ modules/402-ingress-nginx/templates/validator/deployment.yaml
@@ -66,7 +66,15 @@ spec:
       {{- include "helm_lib_tolerations" (tuple $context "any-node") | nindent 6 }}
       {{- include "helm_lib_priority_class" (tuple $context "system-cluster-critical") | nindent 6 }}
       {{- include "helm_lib_pod_affinity" (list $context (dict "app" "validator" "name" $name ) (list "amd64")) | nindent 6 }}
-      {{ include "helm_lib_module_pod_security_context_run_as_user_deckhouse" $context | nindent 6 }}
+      securityContext:
+        runAsNonRoot: true
+        runAsUser: 64535
+        runAsGroup: 64535
+      {{- if eq $validationIsolationMode "IsolatedProcess" }}
+        sysctls:
+        - name: net.ipv4.ip_unprivileged_port_start
+          value: "0"
+      {{- end }}
       serviceAccountName: {{ $validatorServiceAccountName }}
       automountServiceAccountToken: true
       terminationGracePeriodSeconds: 420
